Reorganize tables file structure
